Change secret to longer one. Problem saved, added TestController for testing endpoints via Postman

+++ src/main/java/com/stat_tracker/security/entity/Role.java
@@ -7,24 +7,25 @@ import jakarta.persistence.*;
 public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Enumerated(EnumType.STRING)
     @Column(length = 20)
     private ERole name;
 
     public Role() {
+
     }
 
     public Role(ERole name) {
         this.name = name;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
